Adding dynamic render of tasks.

+++ srt/js/data/tasks.js
@@ -2,9 +2,9 @@ import {folder} from '../model/folder.js'
 
 
 //task generator
-function createTasks(name, tag, id = Math.random()) {
-    folder[name] = [tag, id]
-    const buttons = ['DESTROY IT!', 'DONE', 'FINISH ME', 'SPANG', 'FATALITY!', 'KILLING IT', 'ONE MORE']
+function createTasks(name, tag, id, element) {
+    folder[id] = [name, tag]
+    const buttons = ['DESTROY IT!', 'DONE', 'FINISH ME', 'SPANG!!!', 'FATALITY!', 'KILLING IT', 'ONE MORE']
 
     const tasksMenu = document.getElementById('tasks-menu')
 
@@ -22,10 +22,12 @@ function createTasks(name, tag, id = Math.random()) {
     tag === '✅' ? task.classList.add('none') : 0;
     task.id = id;
     taskText.classList.add('tasks-menu-title')
+    taskTextInput.classList.add('tasks-menu-input')
     taskTag.classList.add('tasks-menu-tag')
     taskButton.classList.add('tasks-menu-button', 'no-select')
     taskButtonDelete.classList.add('tasks-button-delete', 'no-select', 'none')
     taskTextInput.setAttribute('spellcheck', 'false')
+    taskTextInput.setAttribute('maxlength', '40')
 
     // text
     taskTextInput.value = name;
@@ -42,7 +44,25 @@ function createTasks(name, tag, id = Math.random()) {
     taskTag.appendChild(taskTagText)
     taskButton.appendChild(taskButtonText)
 
-    tasksMenu.appendChild(task)
+    element ? element.after(task) : tasksMenu.appendChild(task);
+}
+
+function saveTasks(name, tag) {
+    folder[Math.random()] = [name, tag]
+}
+
+saveTasks('this', '')
+
+function renderTasks(element) {
+    const tasksMenu = document.getElementById('tasks-menu')
+    tasksMenu.innerHTML = ''
+    const entries = Object.entries(folder)
+    entries.forEach(entrie => {
+        const id = entrie[0]
+        const name = entrie[1][0]
+        const tag = entrie[1][1]
+        element ? createTasks(name, tag, id, element) : createTasks(name, tag, id);
+    })
 }
 
 function deleteTasks(id) {
@@ -73,25 +93,27 @@ function toggleNone(element) {
 
 // listeners
 function taskSideBarListener() {
-    const taskSideBar = document.getElementById('tasks-sidebar')
-    taskSideBar.addEventListener('click', e => {
-        const targetClasses = e.target.classList;
-        let result = [];
-        targetClasses.forEach(clase => result.push(clase));
-        const tag = result[1].slice(4);
-        filterTasks(tag);
+    document.addEventListener('click', e => {
+        const taskSide = e.target.parentElement.parentElement;
+        if (taskSide.id === 'tasks-sidebar') {
+            const targetClasses = e.target.classList;
+            let result = [];
+            targetClasses.forEach(clase => result.push(clase));
+            const tag = result[1].slice(4);
+            filterTasks(tag);
+        }
     })
 }
 function deleteButtonListener() {
-    const deleteButton = document.querySelectorAll('.tasks-button-delete')
-    deleteButton.forEach(del => {
-        del.addEventListener('click', e => {
+    document.addEventListener('click', e => {
+        if (e.target.classList.contains('tasks-button-delete')) {
             const currentTaskId = e.target.parentNode.id;
             deleteTasks(currentTaskId)
-        })
+        }
     })
 }
 function taskHoverListener() {
+    // document.addEventListener('mouseover', () => {
     const taskBar = document.querySelectorAll('.tasks-menu-container')
     taskBar.forEach(taskk => {
         taskk.addEventListener('mouseenter', t => {
@@ -101,9 +123,25 @@ function taskHoverListener() {
             toggleNone(t.target.children[2])
         })
     })
+    // })
+}
+
+function taskInputListener() {
+    const allTaskInputs = document.querySelectorAll('.tasks-menu-input');
+    allTaskInputs.forEach(input => {
+        input.addEventListener('keydown', e => {
+            e.key === 'Enter' ? renderTasks(input.parentElement.parentElement) : 0;
+        })
+    })
+}
+
+function allListeners() {
+    taskSideBarListener();
+    deleteButtonListener();
+    taskHoverListener();
+    taskInputListener();
 }
 
 
-
-export {createTasks, deleteTasks, filterTasks, taskSideBarListener, deleteButtonListener, taskHoverListener}
+export {createTasks, deleteTasks, filterTasks, allListeners, renderTasks}
